Add files via upload
Problem set of week 02. The problem sets shortly described:
In readability, the task was to implement a program that computes the approximate grade level needed to comprehend some text using the Coleman-Liau index to calculate this grade of the text. 

In caesar, the task was to implement a program that encrypts messages using a technique called Caesar’s cipher. The Caesar’s algorithm encrypts messages by “rotating” each letter by k positions. So for example if the key is 1 and the inputted message is "HELLO", the program should give as output "IFMMP". So here, each letter of "HELLO" was shifted by 1 in their alphabetical order.